@@ -8,7 +8,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php":">=5.3",
+        "php":">=5.4",
         "ext-openssl": "*"
     },
     "authors": [
