@@ -21,14 +21,14 @@ class Error {
     {
 
         // 本SDK自定义错误类型
-        $code_arr = array(
+        $code_arr = [
             // 错误: get方式请求api网络错误
-            'ERR_GET' => array(13001, 'http get api error.'),
+            'ERR_GET' => [13001, 'http get api error.'],
             // 错误: post方式请求api网络错误
-            'ERR_POST' => array(13002, 'http post api error.'),
+            'ERR_POST' => [13002, 'http post api error.'],
             // 错误: 消息类型未定义
-            'ERR_MEG_TYPE' => array(13003, 'message type is not defined.')
-        );
+            'ERR_MEG_TYPE' => [13003, 'message type is not defined.'],
+        ];
 
         return array((object)array(
                 'errcode' => $code_arr[$code][0],
