@@ -109,10 +109,10 @@ class Wechat
              *
              * Examples:
              * ```
-             * $wechat->reply(array(
+             * $wechat->reply([
              *    'type' => 'text',
              *    'content' => '嘿嘿，呵呵~~'
-             * ));
+             * ]);
              * ```
              */
             case 'text':
@@ -134,10 +134,10 @@ class Wechat
              *
              * Examples:
              * ```
-             * $wechat->reply(array(
-              *     'type' => 'image',
-              *    'media_id' => 'Uq7OczuEGEyUu--dYjg7seTm-EJTa0Zj7UDP9zUGNkVpjcEHhl7tU2Mv8mFRiLKC'
-             * ));
+             * $wechat->reply([
+             *      'type' => 'image',
+             *      'media_id' => 'Uq7OczuEGEyUu--dYjg7seTm-EJTa0Zj7UDP9zUGNkVpjcEHhl7tU2Mv8mFRiLKC'
+             * ]);
              * ```
              */
             case 'image':
@@ -159,10 +159,10 @@ class Wechat
              *
              * Examples:
              * ```
-             * $wechat->reply(array(
-              *    'type' => 'voice',
-              *    'media_id' => 'rVT43tfDwjh4p1BV2gJ5D7Zl2BswChO5L_llmlphLaTPytcGcguBAEJ1qK4cg4r_'
-             * ));
+             * $wechat->reply([
+             *    'type' => 'voice',
+             *    'media_id' => 'rVT43tfDwjh4p1BV2gJ5D7Zl2BswChO5L_llmlphLaTPytcGcguBAEJ1qK4cg4r_'
+             * ]);
              * ```
              */
             case 'voice':
@@ -184,12 +184,12 @@ class Wechat
              *
              * Examples:
              * ```
-             * $wechat->reply(array(
-              *    'type' => 'video',
-              *    'media_id' => 'yV0l71NL0wtpRA8OMX0-dBRQsMVyt3fspPUzurIS3psi6eWOrb_WlEeO39jasoZ8',
-              *    'title' => '视频消息的标题',        //可选
-              *    'description' => '视频消息的描述'   //可选
-             * ));
+             * $wechat->reply([
+            *    'type' => 'video',
+            *    'media_id' => 'yV0l71NL0wtpRA8OMX0-dBRQsMVyt3fspPUzurIS3psi6eWOrb_WlEeO39jasoZ8',
+            *    'title' => '视频消息的标题',        //可选
+            *    'description' => '视频消息的描述'   //可选
+             * ]);
              * ```
              */
             case 'video':
@@ -217,14 +217,14 @@ class Wechat
              *
              * Examples:
              * ```
-             * $wechat->reply(array(
+             * $wechat->reply([
               *    'type' => 'music',
               *    'title' => '音乐标题',           //可选
               *    'description' => '音乐描述',     //可选
               *    'music_url' => 'http://me.diary8.com/data/music/2.mp3',      //可选
               *    'hqmusic_url' => 'http://me.diary8.com/data/music/2.mp3',    //可选
               *    'thumb_media_id' => 'O39wW0ZsXCb5VhFoCgibQs5PupFb6VZ2jH5A8gHUJCJz2Qmkrb7objoTue7bGTGQ',
-             * ));
+             * ]);
              * ```
              */
             case 'music':
@@ -256,29 +256,29 @@ class Wechat
              *
              * Examples:
              * ```
-             * $wechat->reply(array(
+             * $wechat->reply([
              *    'type' => 'news',
-             *    'articles' => array(
-             *        array(
+             *    'articles' => [
+             *        [
              *            'title' => '图文消息标题1',                                //可选
              *            'description' => '图文消息描述1',                          //可选
              *            'picurl' => 'http://me.diary8.com/data/img/demo1.jpg',   //可选
              *            'url' => 'http://www.example.com/'                       //可选
-             *        ),
-             *        array(
+             *        ],
+             *        [
              *            'title' => '图文消息标题2',
              *            'description' => '图文消息描述2',
              *            'picurl' => 'http://me.diary8.com/data/img/demo2.jpg',
              *            'url' => 'http://www.example.com/'
-             *        ),
-             *        array(
+             *        ],
+             *        [
              *            'title' => '图文消息标题3',
              *            'description' => '图文消息描述3',
              *            'picurl' => 'http://me.diary8.com/data/img/demo3.jpg',
              *            'url' => 'http://www.example.com/'
-             *        )
-             *     )
-             * ));
+             *        ],
+             *     ],
+             * ]);
              * ```
              */
             case 'news':
@@ -315,10 +315,10 @@ class Wechat
              *
              * Examples:
              * ```
-             * $wechat->reply(array(
+             * $wechat->reply([
              *    'type' => 'transfer_customer_service',
              *    'kf_account' => 'test1@test'            // 可选
-             * ));
+             * ]);
              * ```
              */
             case 'transfer_customer_service':
@@ -346,13 +346,13 @@ class Wechat
              *
              * Examples:
              * ```
-             * $wechat->reply(array(
+             * $wechat->reply([
               *    'type' => 'device_text',
               *    'device_type' => '设备类型，目前为“公众账号原始ID”',
               *    'device_id' => '设备ID，第三方提供',
               *    'session_id' => '微信客户端生成的session id',
               *    'content' => '消息内容，BASE64编码'
-             * ));
+             * ]);
              * ```
              */
             case 'device_text':
@@ -380,12 +380,12 @@ class Wechat
              *
              * Examples:
              * ```
-             * $wechat->reply(array(
+             * $wechat->reply([
               *    'type' => 'device_status',
               *    'device_type' => '设备类型，目前为“公众账号原始ID”',
               *    'device_id' => '设备ID，第三方提供',
               *    'deviceStatus' => '设备状态:0未连接,1已连接'
-             * ));
+             * ]);
              * ```
              */
             case 'device_status':
